feat: migrate to 1:1 Minecraft 1.12.2 block/item system, sync starting equipment & creative inventory, and implement dynamic Chinese localization

+++ src/systems/SaveSystem.ts
@@ -36,7 +36,7 @@ export interface SaveData {
     armor: ({ id: number; count: number } | null)[];
   };
   seed: number;
-  chunks: { cx: number; cz: number; data: Uint8Array; metadata?: SerializedBlockMetadata[] }[];
+  chunks: { cx: number; cz: number; data: Uint16Array; metadata?: SerializedBlockMetadata[] }[];
   timestamp: number;
 }
